new preferences system working, added multilanguage support, screen transitions, weapon movement fixed and new weapon ai with improve cooldown

+++ core/src/main/kotlin/com/gabriel/component/SkillComponent.kt
@@ -10,13 +10,13 @@ enum class Skill(
     val skillEntityId: Int,
     var skillName: String = "",
     var skillLevel: Int = 0,
-    var onLevelUP: Int,
+    var onLevelUP: Float,
 ) {
-    PLAYER_LIFE(0, "Life", 0, 100),
-    PLAYER_REGENERATION(1, "Regeneration", 0, 1),
-    PLAYER_SPEED(2, "Speed", 0, 1),
-    PLAYER_COOLDOWN(3, "Cooldown", 0, -1),
-    PLAYER_DAMAGE(4, "Damage", 0, 1);
+    PLAYER_LIFE(0, "Life", 0, 100f),
+    PLAYER_REGENERATION(1, "Regeneration", 0, 1f),
+    PLAYER_SPEED(2, "Speed", 0, 1f),
+    PLAYER_COOLDOWN(3, "Cooldown", 0, -1f),
+    PLAYER_DAMAGE(4, "Damage", 0, 1f);
 
     var atlasKey: String = this.toString().lowercase()
 }
